ecommerce/requirements.txt
Flask==2.0.1
grpcio==1.38.1
grpcio-tools==1.38.1
pymongo==3.11.4
